(bug) Uncomment wsus initiali syncrhonization exec
During development i commented out wsus's initial synchronization task.  Unfortunately when users use this module this causes runs to say changes need to happen, attempt to be performed, but fail to truly happen but now blow up under the covers.  This causes the next run to do the same so puppet keeps attempting to converge the system until the initial sync is performed by the user.  This should resolve that issue.