wsusserver_install: Remove DSC dependency (#54)
* Remove DSC dependency

The DSC module is no longer recommended by Puppet. It requires syncing thousands of files to every computer that checks into a given Puppet environment. Thankfully its use was rather trivial here and can easily be replaced by the dedicated windowsfeature module. Thanks to Kevin Reeuwijk for initially raising this issue and putting in a PR.

* Update tests to use windowsfeature instead of DSC